% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certedata.R
\docType{package}
\name{certedata-package}
\alias{certedata}
\alias{certedata-package}
\title{certedata: A Certe R Package for Loading Certe Universe}
\description{
A Certe R Package for loading all R packages in the Certe
  Medical Epidemiology universe, based on the tidyverse/tidyverse repo.
  This package is part of the 'certedata' universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://certe-medical-epidemiology.github.io/certedata}
  \item \url{https://github.com/certe-medical-epidemiology/certedata}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.berends@certe.nl}

Authors:
\itemize{
  \item Erwin E. A. Hassing
}

Other contributors:
\itemize{
  \item Certe Medical Diagnostics & Advice Foundation [copyright holder, funder]
}

}
\keyword{internal}
