% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{certedata_attach}
\alias{certedata_attach}
\title{Attach 'certedata' Universe Packages}
\usage{
certedata_attach(...)
}
\arguments{
\item{...}{unused at the moment}
}
\description{
Attach 'certedata' Universe Packages
}
\details{
This function attaches all 'certedata' universe R packages and their accompanying third-party packages (in total 26 packages).

To install the packages, see \code{\link[=certedata_install_packages]{certedata_install_packages()}}.
}
