% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{certedata_deps}
\alias{certedata_deps}
\title{List all certedata dependencies}
\usage{
certedata_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
certedata packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all certedata dependencies
}
