% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{certedata_install_packages}
\alias{certedata_install_packages}
\title{Install all packages required for the 'certedata' universe}
\usage{
certedata_install_packages()
}
\description{
This function not only installs the 'certedata' universe, but also its accompanying third-party packages.
}
\details{
These 25 packages will be installed if not already:
\itemize{
\item \code{AMR}
\item \code{broom}
\item \code{certedb}
\item \code{certegis}
\item \code{certemail}
\item \code{certeplot2}
\item \code{certeprojects}
\item \code{certestats}
\item \code{certestyle}
\item \code{certetoolbox}
\item \code{cleaner}
\item \code{dplyr}
\item \code{extrafont}
\item \code{flextable}
\item \code{forcats}
\item \code{ggplot2}
\item \code{glue}
\item \code{lubridate}
\item \code{purrr}
\item \code{readr}
\item \code{readxl}
\item \code{skimr}
\item \code{stringr}
\item \code{tibble}
\item \code{tidyr}
}
}
