% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{certedata_packages}
\alias{certedata_packages}
\title{List all packages in the certedata}
\usage{
certedata_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include certedata in the list?}
}
\description{
List all packages in the certedata
}
\examples{
certedata_packages()
}
