% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{certedata_packages}
\alias{certedata_packages}
\title{List all packages in the 'certedata' universe}
\usage{
certedata_packages(only_installed = TRUE, include_self = TRUE)
}
\arguments{
\item{only_installed}{a \link{logical} to indicate whether only currently installed packages should be returned}

\item{include_self}{a \link{logical} to indicate whether the 'certedata' package should be included as well}
}
\description{
List all packages in the 'certedata' universe
}
\examples{
certedata_packages()
}
